requirements.txt
# Core dependencies
google-ads==26.1.0
grpcio>=1.59.0,<2.0.0
google-api-core>=2.13.0,<=3.0.0
PyYAML>=5.1,<7.0
google-auth-oauthlib>=1.0.0,<2.0.0
protobuf>=4.25.0,<7.0.0
proto-plus>=1.22.3,<2.0.0
googleapis-common-protos>=1.56.3,<2.0.0
grpcio-status>=1.59.0,<2.0.0
requests>=2.18.0,<3.0.0
google-auth>=2.14.1,<3.0.0

# Agent dependencies
python-dotenv>=1.0.0
google-generativeai>=0.3.0  # For Gemini model
zoneinfo>=0.2.1  # For timezone handling
pydantic>=2.0.0  # For data validation
beautifulsoup4>=4.13.0  # For HTML parsing

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0 